TM-68 - Changed logic to send to checkout per tony
The default behavior of the checkout page should be to start new order.
The switch ?task=startneworder should not be needed since this is the default assumption if no active order is present.
When on upsell page 1 or 2 and more than 58min has passed since the order is created, forward to the thankyou page, since the upsells are no longer possible.